Move verifier code around so it's cleaner.
